♻️ move getZoneJsOriginalValue to @datadog/js-core/util
Move `getZoneJsOriginalValue` and `BrowserWindowWithZoneJs` from
`browser-core/src/tools/getZoneJsOriginalValue` to
`js-core/src/util/getZoneJsOriginalValue`. Marked as `@internal`
in the js-core API report — not part of the supported public API.

The spec stays in browser-core (it depends on browser-specific test
helpers like `mockZoneJs`).

All consumers (browser-core internals: fetch, addEventListener,
timer; browser-rum-core: domMutationObservable) now import directly
from `@datadog/js-core/util`.
