+++ packages/browser-core/src/browser/addEventListener.ts
@@ -1,6 +1,6 @@
 import type { CookieStore, CookieStoreEventMap } from '@datadog/js-core/util'
 import { monitor } from '@datadog/js-core/monitor'
-import { getZoneJsOriginalValue } from '../tools/getZoneJsOriginalValue'
+import { getZoneJsOriginalValue } from '@datadog/js-core/util'
 import { noop } from '../tools/utils/functionUtils'
 import type { VisualViewport, VisualViewportEventMap } from './browser.types'
