+++ packages/js-core/src/entries/util.ts
@@ -39,3 +39,5 @@ export type { Context, ContextValue, ContextArray } from '../util/context'
 export { mockable, mockableReplacements } from '../util/mockable'
 export { jsonStringify, detachToJsonMethod } from '../util/jsonStringify'
 export type { ObjectWithToJsonMethod } from '../util/jsonStringify'
+export { getZoneJsOriginalValue } from '../util/getZoneJsOriginalValue'
+export type { BrowserWindowWithZoneJs } from '../util/getZoneJsOriginalValue'
